requirements.txt
anyio==4.4.0
APScheduler==3.10.4
attrs==24.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
selenium==4.23.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.4.1
trio==0.26.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
